+++ src/index.js
@@ -33,7 +33,6 @@ const App = () => {
           </CSSTransition>
         </TransitionGroup>
       )} />
-      <Footer />
     </ >
   )
 };
